move metaclasses to meta.py
