+++ pydec/nn/modules/__init__.py
@@ -8,3 +8,4 @@ from .container import (
     ParameterList,
     ParameterDict,
 )
+from . import meta
